requirements.txt
camelot-py~=0.11.0
pandas~=2.2.3
PyYAML~=6.0.1
ghostscript~=0.7